MAT-8026 added SearchUtils class and some updated tests
